Cleaning code to make tests works

+++ gegelatilib/src/file/tpgGraphDotImporter.cpp
@@ -188,7 +188,7 @@ void File::TPGGraphDotImporter::readProgram(std::smatch& matches)
         for (;;) {
             if (pos1 != std::string::npos) {
                 v_constant.push_back(
-                    {std::stoi(this->lastLine.substr(pos, pos1 - pos))});
+                    {std::stod(this->lastLine.substr(pos, pos1 - pos))});
             }
             else {
                 break;
