+++ test/exporterTest.cpp
@@ -99,7 +99,7 @@ class ExporterTest : public ::testing::Test
                 std::make_shared<Program::Program>(*e, false);
             for (int j = 0; j < constant_size; j++) {
                 p.get()->getConstantHandler().setDataAt(typeid(Data::Constant),
-                                                        j, {j - 2});
+                                                        j, {(double)(j - 2)});
             }
             progPointers.push_back(p);
         }
